# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wander REST API",
    "version": "1.0.0",
    "description": "Stores trips for their owners. Every write carries the trip `version` it was based on and\nfails with `409` when the stored trip has moved on.\n\n## Authorization\n\nRoutes marked with a lock require `Authorization: Bearer <token>` and answer `401` without\nit. Select **Authorize** and enter the token alone, without the `Bearer` prefix.\n\n- `APP_MODE=demo`: the token is `demo`. This page enters it for you.\n- `APP_MODE=live`: use a Cognito access token issued to the configured app client.\n\nRoutes without a lock are public: the shared-trip projection and the health check."
  },
  "servers": [{ "url": "/" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The token alone, without the Bearer prefix. In demo mode it is `demo`."
      }
    }
  },
  "paths": {
    "/trips": {
      "get": {
        "summary": "List owned trips",
        "tags": ["Trips"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "ownerId": { "type": "string", "minLength": 1 },
                          "city": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 100
                          },
                          "preferences": {
                            "type": "object",
                            "properties": {
                              "interests": {
                                "type": "array",
                                "items": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 80
                                },
                                "maxItems": 20
                              },
                              "pace": {
                                "type": "string",
                                "enum": ["relaxed", "balanced", "busy"]
                              }
                            },
                            "required": ["interests", "pace"],
                            "additionalProperties": false
                          },
                          "days": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": { "type": "string", "format": "uuid" },
                                "date": { "type": "string", "format": "date" },
                                "activities": {
                                  "type": "array",
                                  "items": {
                                    "type": "object",
                                    "properties": {
                                      "id": {
                                        "type": "string",
                                        "format": "uuid"
                                      },
                                      "title": {
                                        "type": "string",
                                        "minLength": 1,
                                        "maxLength": 120
                                      },
                                      "pinned": { "type": "boolean" }
                                    },
                                    "required": ["id", "title", "pinned"],
                                    "additionalProperties": false
                                  },
                                  "maxItems": 20
                                }
                              },
                              "required": ["id", "date", "activities"],
                              "additionalProperties": false
                            },
                            "minItems": 1,
                            "maxItems": 30
                          },
                          "share": {
                            "type": "object",
                            "properties": {
                              "token": { "type": "string", "minLength": 32 },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "expiresAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": ["token", "createdAt", "expiresAt"],
                            "additionalProperties": false,
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "version": {
                            "type": "integer",
                            "exclusiveMinimum": true,
                            "minimum": 0
                          }
                        },
                        "required": [
                          "id",
                          "ownerId",
                          "city",
                          "preferences",
                          "days",
                          "share",
                          "createdAt",
                          "updatedAt",
                          "version"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": { "type": "string", "nullable": true }
                  },
                  "required": ["items", "nextCursor"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": { "description": "Invalid request" },
          "401": { "description": "Missing or invalid bearer token" },
          "404": { "description": "Not found" },
          "409": { "description": "Version conflict" },
          "429": { "description": "Rate limit exceeded" }
        },
        "parameters": [
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100 }
          }
        ]
      },
      "post": {
        "summary": "Create trip",
        "tags": ["Trips"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "city": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "preferences": {
                    "type": "object",
                    "properties": {
                      "interests": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 80
                        },
                        "maxItems": 20
                      },
                      "pace": {
                        "type": "string",
                        "enum": ["relaxed", "balanced", "busy"]
                      }
                    },
                    "required": ["interests", "pace"],
                    "additionalProperties": false
                  },
                  "days": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "date": { "type": "string", "format": "date" },
                        "activities": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string", "format": "uuid" },
                              "title": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 120
                              },
                              "pinned": { "type": "boolean" }
                            },
                            "required": ["id", "title", "pinned"],
                            "additionalProperties": false
                          },
                          "maxItems": 20
                        }
                      },
                      "required": ["id", "date", "activities"],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 30
                  }
                },
                "required": ["city", "preferences", "days"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "ownerId": { "type": "string", "minLength": 1 },
                    "city": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "preferences": {
                      "type": "object",
                      "properties": {
                        "interests": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 80
                          },
                          "maxItems": 20
                        },
                        "pace": {
                          "type": "string",
                          "enum": ["relaxed", "balanced", "busy"]
                        }
                      },
                      "required": ["interests", "pace"],
                      "additionalProperties": false
                    },
                    "days": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "date": { "type": "string", "format": "date" },
                          "activities": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": { "type": "string", "format": "uuid" },
                                "title": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 120
                                },
                                "pinned": { "type": "boolean" }
                              },
                              "required": ["id", "title", "pinned"],
                              "additionalProperties": false
                            },
                            "maxItems": 20
                          }
                        },
                        "required": ["id", "date", "activities"],
                        "additionalProperties": false
                      },
                      "minItems": 1,
                      "maxItems": 30
                    },
                    "share": {
                      "type": "object",
                      "properties": {
                        "token": { "type": "string", "minLength": 32 },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "expiresAt": { "type": "string", "format": "date-time" }
                      },
                      "required": ["token", "createdAt", "expiresAt"],
                      "additionalProperties": false,
                      "nullable": true
                    },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "updatedAt": { "type": "string", "format": "date-time" },
                    "version": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "minimum": 0
                    }
                  },
                  "required": [
                    "id",
                    "ownerId",
                    "city",
                    "preferences",
                    "days",
                    "share",
                    "createdAt",
                    "updatedAt",
                    "version"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": { "description": "Invalid request" },
          "401": { "description": "Missing or invalid bearer token" },
          "404": { "description": "Not found" },
          "409": { "description": "Version conflict" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
...
    "/shared/{ownerId}/{tripId}/{token}": {
      "get": {
        "summary": "Read public share",
        "tags": ["Sharing"],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "city": { "type": "string" },
                    "days": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "date": { "type": "string", "format": "date" },
                          "activities": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": { "type": "string", "format": "uuid" },
                                "title": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 120
                                },
                                "pinned": { "type": "boolean" }
                              },
                              "required": ["id", "title", "pinned"],
                              "additionalProperties": false
                            },
                            "maxItems": 20
                          }
                        },
                        "required": ["id", "date", "activities"],
                        "additionalProperties": false
                      }
                    },
                    "updatedAt": { "type": "string", "format": "date-time" }
                  },
                  "required": ["id", "city", "days", "updatedAt"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": { "description": "Invalid request" },
          "404": { "description": "Not found" },
          "429": { "description": "Rate limit exceeded" }
        },
        "parameters": [
          {
            "name": "ownerId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ]
      }
    },
...
  }
}